requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
psycopg2-binary==2.9.3
pytest-factoryboy==2.1.0
django-filter==21.1
django-extensions==3.1.5
pytest-clarity==1.0.1
pytest-cov==3.0.0
pytest-icdiff==0.5